+++ modules/framegrabber/include/o3d3xx_framegrabber/pcic_schema.h
@@ -31,6 +31,7 @@ namespace o3d3xx
   extern const std::uint16_t IMG_RAMP; // raw amplitude
   extern const std::uint16_t IMG_CART; // Cartesian
   extern const std::uint16_t IMG_UVEC; // Unit vectors
+  extern const std::uint16_t EXP_TIME; // Exposure times
 
   /**
    * Utility function to build a PCIC schema from a mask